+++ examples/showcase/.storybook/main.js
@@ -11,7 +11,7 @@ const config = {
     '@chromatic-com/storybook',
   ],
   framework: {
-    name: '@storybook-vue/nuxt',
+    name: '@tian-zhihui/storybook-vue-nuxt',
     options: {},
   },
   docs: {